AI_Module/requirements.txt
fastapi
uvicorn
python-dotenv
requests
pydantic
pytest
httpx
google-generativeai